third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-import-block-but-unsafe-inline-nested-expected.txt

PASS document.inlineScriptHasRun is true
PASS document.externalScriptHasRun is undefined
PASS document.corsExternalScriptHasRun is undefined
FAIL document.evalFromInlineHasRun should be undefined (of type undefined). Was true (of type boolean).
PASS document.evalFromExternalHasRun is undefined
PASS document.evalFromCorsExternalHasRun is undefined
PASS successfullyParsed is true

TEST COMPLETE
